pyproject.toml
[project]
name = "card-game"
version = "0.1.0"
description = "卡牌對戰機器人"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "alembic-postgresql-enum>=1.8.0",
    "asyncpg>=0.30.0",
    "authlib>=1.6.5",
    "cryptography>=46.0.3",
    "discord-py>=2.6.4",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "jishaku>=2.6.3",
    "loguru>=0.7.3",
    "openai>=2.7.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
authors = [{ "name" = "seriaati", "email" = "seria.ati@gmail.com" }]

[tool.pyright]
enableTypeIgnoreComments = false
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportUnnecessaryIsInstance = true
reportUnnecessaryTypeIgnoreComment = true
typeCheckingMode = "standard"
include = ["app", "run.py", "tests", "bot"]
exclude = ["alembic/versions"]

[dependency-groups]
type-check = ["pyright>=1.1.407"]
